requirements.txt
# Windows网络流量智能分析工具 - 依赖清单

# 核心依赖
#python>=3.10

# JSONL处理
jsonlines>=3.1.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 配置文件处理
pyyaml>=6.0

# 时间处理
python-dateutil>=2.8.2

# HTTP客户端 (用于AI API调用)
requests>=2.28.0

# 文件编码检测
chardet>=5.0.0

# 可选依赖 (用于Web界面，v2.0)
# streamlit>=1.28.0
# fastapi>=0.100.0
# uvicorn>=0.23.0

# 开发依赖 (可选)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
